[UT] Refactor shared-data ut code for warehouse related cases
