+++ fe/fe-core/src/test/java/com/starrocks/catalog/GlobalStateMgrTestUtil.java
@@ -49,6 +49,7 @@ import com.starrocks.sql.ast.SingleRangePartitionDesc;
 import com.starrocks.system.Backend;
 import com.starrocks.thrift.TStorageMedium;
 import com.starrocks.thrift.TStorageType;
+import com.starrocks.utframe.UtFrameUtils;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
@@ -113,6 +114,9 @@ public class GlobalStateMgrTestUtil {
                 testStartVersion);
         LocalMetastore metastore = (LocalMetastore) globalStateMgr.getLocalMetastore();
         metastore.unprotectCreateDb(db);
+
+        // init default warehouse
+        UtFrameUtils.mockInitWarehouseEnv();
         return globalStateMgr;
     }
