+++ fe/fe-core/src/test/java/com/starrocks/scheduler/TaskBuilderTest.java
@@ -16,11 +16,7 @@ package com.starrocks.scheduler;
 
 import com.starrocks.catalog.MaterializedView;
 import com.starrocks.catalog.TableProperty;
-import com.starrocks.server.WarehouseManager;
-import com.starrocks.warehouse.DefaultWarehouse;
-import com.starrocks.warehouse.Warehouse;
-import mockit.Mock;
-import mockit.MockUp;
+import com.starrocks.utframe.UtFrameUtils;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -29,19 +25,7 @@ public class TaskBuilderTest {
     @Test
     public void testTaskBuilderForMv() {
         // mock the warehouse of MaterializedView for creating task
-        new MockUp<WarehouseManager>() {
-            @Mock
-            public Warehouse getWarehouse(long warehouseId) {
-                return new DefaultWarehouse(WarehouseManager.DEFAULT_WAREHOUSE_ID,
-                        WarehouseManager.DEFAULT_WAREHOUSE_NAME);
-            }
-
-            @Mock
-            public Warehouse getWarehouse(String warehouse) {
-                return new DefaultWarehouse(WarehouseManager.DEFAULT_WAREHOUSE_ID,
-                        WarehouseManager.DEFAULT_WAREHOUSE_NAME);
-            }
-        };
+        UtFrameUtils.mockInitWarehouseEnv();
 
         MaterializedView mv = new MaterializedView();
         mv.setName("aa.bb.cc");
